Moved ParameterDummy from multiple Test classes into SequencingDummies collection.
